+++ src/invariants.ts
@@ -1,39 +1,10 @@
-import { TotalRO, OrderRO, ShippingRO, ItemsRO, ItemTotalRO, ItemQtyRO } from './readonlyInterfaces'
 import { minusPrice } from './basics'
-import { documents as _ } from './documents'
+import { documents } from './documents'
+import { invariants as i } from './injectable/invariants'
 
 export const invariants = {
-    total: <T extends OrderRO<TotalRO>>({ total, invoiced, canceled, refunded }: T) => ({
-        ci: minusPrice(total, _.total([...canceled, ...invoiced])),
-        ir: minusPrice(_.total(invoiced), _.total(refunded)) }),
-    shipping: <T extends OrderRO<ShippingRO>>({ shipping, invoiced, refunded, canceled }: T) => ({
-        total: ({
-            ci: minusPrice(shipping, _.shipping([...canceled, ...invoiced])),
-            ir: minusPrice(_.shipping(invoiced), _.shipping(refunded)) }),
-        qty: (([invoiced, refunded, canceled]) => (wrong => wrong.length !== 0
-                ? 1 - wrong[0].length
-                : canceled.length === 0 || invoiced.length === 0
-                    ? invoiced.length - refunded.length
-                    : -canceled.length
-            )([invoiced, refunded, canceled].filter(documents => documents.length > 1))
-        )([invoiced, refunded, canceled]
-            .map(documents => documents.filter(({ shipping }) => shipping !== 0))) }),
+    total: i.total({ mp: minusPrice, dt: documents.total }),
+    shipping: i.shipping({ mp: minusPrice, ds: documents.shipping }),
     items: {
-        qty: <V extends ItemQtyRO, U extends ItemsRO<V>, T extends OrderRO<U>>(order: T) => (reducer => ({
-            ci: reducer([order], [...order.canceled, ...order.invoiced]),
-            ir: reducer(order.invoiced, order.refunded)
-        }))((from: readonly U[], subtrahend: readonly U[]) => _.items.qty<V, U>([
-            ...from,
-            ...subtrahend.map(document => ({
-                ...document,
-                items: document.items.map(({ qty, ...item }) => ({ ...item, qty: -qty })) }))])),
-       total: <V extends ItemTotalRO, U extends ItemsRO<V>, T extends OrderRO<U>>(order: T) => (reducer => ({
-            ci: reducer([order], [...order.canceled, ...order.invoiced]),
-            ir: reducer(order.invoiced, order.refunded)
-        }))((from: readonly U[], subtrahend: readonly U[]) => _.items.total<V, U>([
-            ...from,
-            ...subtrahend.map(document => ({
-                ...document,
-                items: document.items.map(({ total, ...item }) => ({ ...item, total: -total })) }))]))
-    }
-}
+        qty: i.items.qty(documents.items.qty),
+        total: i.items.total(documents.items.total) } }