+++ src/tests/sales/sales.ts
@@ -14,13 +14,14 @@ export const checkDocument = <T extends ResultDocument & { items: (CartItem & To
         expect(new Set(data.items)).toEqual(new Set(items)) }
 
 export const scenario = <T extends (cart: C) => C & Items<U & Total> & Total, U extends CartItem, C extends Cart<U>>(totals: T) =>
-    (cart: C) => (order => ({
-        result: order,
-        invoice: doInvoice(order),
-        refund: doRefund(order),
-        cancel: doCancel(order),
-        spreading: c.order<U & Total, Order<Shipping & Total & Items<U & Total>>, Cart<U & Total> & C>(order),
-    }))(deepFreeze({ ...cart, ...totals(cart), canceled: [], invoiced: [], refunded: [] }) as Order<Shipping & Total & Items<U & Total>>)
+    (cart: C) => {
+        const order: Order<Shipping & Total & Items<U & Total>> = deepFreeze({ ...cart, ...totals(cart), canceled: [], invoiced: [], refunded: [] })
+        return { result: order,
+            invoice: doInvoice(order),
+            refund: doRefund(order),
+            cancel: doCancel(order),
+            spreading: c.order<U & Total, Order<Shipping & Total & Items<U & Total>>>(order) }
+    }
 
 const doInvoice = <
     S extends U & Total,
@@ -28,16 +29,17 @@ const doInvoice = <
     U extends CartItem,
     C extends Cart<U>,
     T extends Order<V>
->(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, invoice: C) => (invoice =>
-(order => ({
-    result: invoice,
-    order,
-    invoice: doInvoice<S, V, U, C, T>(order),
-    refund: doRefund<S, V, U, C, T>(order),
-    cancel: doCancel<S, V, U, C, T>(order),
-    spreading: c.order<S, T, Cart<S>>(order),
-}))(deepFreeze({ ...order, invoiced: [...order.invoiced, invoice] }))
-)((cart => cart.total(rulesCalculator(cart)))(o.total<U, C, S, V, T>(order).invoice(deepFreeze(invoice))))
+>(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, invoice: C) => {
+    const cart_ = o.total<U, S, V, T>(order).invoice(deepFreeze(invoice))
+    const invoice_ = cart_.total(rulesCalculator(cart_))
+    const order_ = deepFreeze({ ...order, invoiced: [...order.invoiced, invoice_] })
+    return { result: invoice_,
+        order: order_,
+        invoice: doInvoice<S, V, U, C, T>(order_),
+        refund: doRefund<S, V, U, C, T>(order_),
+        cancel: doCancel<S, V, U, C, T>(order_),
+        spreading: c.order<S, T>(order_) }
+}
 
 const doRefund = <
     S extends U & Total,
@@ -45,16 +47,17 @@ const doRefund = <
     U extends CartItem,
     C extends Cart<U>,
     T extends Order<V>
->(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, refund: C) => (refund =>
-(order => ({
-    result: refund,
-    order,
-    invoice: doInvoice<S, V, U, C, T>(order),
-    refund: doRefund<S, V, U, C, T>(order),
-    cancel: doCancel<S, V, U, C, T>(order),
-    spreading: c.order<S, T, Cart<S>>(order),
-}))(deepFreeze({ ...order, refunded: [...order.refunded, refund] }))
-)((cart => cart.total(rulesCalculator(cart)))(o.total<U, C, S, V, T>(order).refund(deepFreeze(refund))))
+>(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, refund: C) => {
+    const cart_ = o.total<U, S, V, T>(order).refund(deepFreeze(refund))
+    const refund_ = cart_.total(rulesCalculator(cart_))
+    const order_ = deepFreeze({ ...order, refunded: [...order.refunded, refund_] })
+    return { result: refund_,
+        order: order_,
+        invoice: doInvoice<S, V, U, C, T>(order_),
+        refund: doRefund<S, V, U, C, T>(order_),
+        cancel: doCancel<S, V, U, C, T>(order_),
+        spreading: c.order<S, T>(order_) }
+}
 
 const doCancel = <
     S extends U & Total,
@@ -62,13 +65,14 @@ const doCancel = <
     U extends CartItem,
     C extends Cart<U>,
     T extends Order<V>
->(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, cancel: C) => (cancel =>
-(order => ({
-    result: cancel,
-    order,
-    invoice: doInvoice<S, V, U, C, T>(order),
-    refund: doRefund<S, V, U, C, T>(order),
-    cancel: doCancel<S, V, U, C, T>(order),
-    spreading: c.order<S, T, Cart<S>>(order),
-}))(deepFreeze({ ...order, canceled: [...order.canceled, cancel] }))
-)((cart => cart.total(rulesCalculator(cart)))(o.total<U, C, S, V, T>(order).cancel(deepFreeze(cancel))))
+>(order: T) => <R extends (cart: C & Items<S>) => V>(rulesCalculator: R, cancel: C) => {
+    const cart_ = o.total<U, S, V, T>(order).cancel(deepFreeze(cancel))
+    const cancel_ = cart_.total(rulesCalculator(cart_))
+    const order_ = deepFreeze({ ...order, canceled: [...order.canceled, cancel_] })
+    return { result: cancel_,
+        order: order_,
+        invoice: doInvoice<S, V, U, C, T>(order_),
+        refund: doRefund<S, V, U, C, T>(order_),
+        cancel: doCancel<S, V, U, C, T>(order_),
+        spreading: c.order<S, T>(order_) }
+}